Confirmação de remover movimentação finalizada

+++ app/src/main/java/com/mycompany/confinance/view/viewholder/ExpenseViewHolder.kt
@@ -1,13 +1,17 @@
 package com.mycompany.confinance.view.viewholder
 
+import android.app.AlertDialog
 import androidx.recyclerview.widget.RecyclerView
 import com.mycompany.confinance.databinding.RowRecyclerBinding
 import com.mycompany.confinance.model.movement.GetMovementModel
 import com.mycompany.confinance.view.OnMovementListener
 
-class ExpenseViewHolder(private val bind: RowRecyclerBinding, private val listener : OnMovementListener): RecyclerView.ViewHolder(bind.root){
+class ExpenseViewHolder(
+    private val bind: RowRecyclerBinding,
+    private val listener: OnMovementListener
+) : RecyclerView.ViewHolder(bind.root) {
 
-    fun bind(movement: GetMovementModel){
+    fun bind(movement: GetMovementModel) {
         bind.textInformation.text = "${movement.description} - ${movement.date}"
         bind.textValue.text = "$ ${movement.value}"
 
@@ -15,8 +19,17 @@ class ExpenseViewHolder(private val bind: RowRecyclerBinding, private val listen
             listener.onClick(movement.id!!)
         }
         bind.viewRecycler.setOnLongClickListener {
-            listener.onDelete(movement.id!!)
+            AlertDialog.Builder(itemView.context)
+           .setTitle("Remoção da movimentação")
+           .setMessage("Você tem certeza que deseja remover?")
+           .setPositiveButton("Sim") { dialog, which ->
+              listener.onDelete(movement.id!!)
+                }
+                .setNegativeButton("Cancelar", null)
+                .create()
+                .show()
+
             true
         }
     }
-    }
+}
